refactor(engine): move the ICE filters and the SDP helpers out of connection.go
connection.go 631 -> 541, into two files named for what they hold: icefilter.go
(keepICEIP, makeInterfaceAllowFilter) and sdp.go (extractFingerprint,
patchMaxMessageSize).

icefilter_test.go already existed and already tested exactly the two functions
icefilter.go now holds. The test layout has been asserting this seam for longer
than the sources have honored it.

Neither import survives the move: net and strings were used only by the code
that left, so connection.go drops both.

No behavior change, and the proof is mechanical rather than asserted. Against a
worktree at the merge base:

  diff <(cd ../floe-base/cli && go doc -all -u ./engine/peer) \
       <(cd cli && go doc -all -u ./engine/peer)

is empty. The -u is the point: it includes unexported symbols and their doc
comments, and go doc sorts within CONSTANTS, FUNCTIONS and TYPES rather than by
file, so a pure move is empty by construction while a dropped or renamed symbol
is not.

That proof earned its place on the first run. Writing each new file as

    // <what this file holds>
    package peer

makes the note a PACKAGE doc comment, and Go concatenates those across files,
so the package documentation silently grew two paragraphs. The diff caught it
before it left the machine. Both notes now sit below the package clause.

Test plan
- go doc -all -u diff against the base worktree: empty.
- All four moved function bodies md5-identical to their base copies.
- go vet and go test clean in cli/ and in desktop/, which links this package.
- gofmt -l names the same three pre-existing CRLF files as the base worktree,
  and neither new file.
